FEATURE: moves state logic into ConfigService
